+++ UnitTests/BraintreeLocalPaymentTests/BTLocalPaymentClient_UnitTests.swift
@@ -28,7 +28,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         client.apiClient = mockAPIClient
         let expectation = expectation(description: "Start payment fails with error")
 
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.fetchConfigurationFailed.errorCode)
@@ -44,7 +44,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         client.apiClient = mockAPIClient
         let expectation = expectation(description: "Start payment fails with error")
 
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.disabled.errorCode)
@@ -62,7 +62,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         let expectation = expectation(description: "Start payment fails with error")
         localPaymentRequest.localPaymentFlowDelegate = nil
 
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.integration.errorCode)
@@ -113,7 +113,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         )
         paymentRequest.localPaymentFlowDelegate = mockLocalPaymentRequestDelegate
         
-        client.startPaymentFlow(paymentRequest) { _, _ in }
+        client.start(paymentRequest) { _, _ in }
 
         XCTAssertEqual(mockAPIClient.lastPOSTParameters!["merchant_account_id"] as? String, "customer-nl-merchant-account")
         XCTAssertEqual(mockAPIClient.lastPOSTParameters!["funding_source"] as? String, "ideal")
@@ -156,7 +156,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         mockAPIClient.cannedResponseBody = BTJSON(value: ["paymentResource": ["paymentToken": "123aaa-123-543-777"]])
         let expectation = expectation(description: "Start payment fails with error")
 
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.appSwitchFailed.errorCode)
@@ -180,7 +180,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         )
         let expectation = expectation(description: "Start payment fails with error")
 
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.appSwitchFailed.errorCode)
@@ -205,7 +205,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         waitForExpectations(timeout: 4)
 
@@ -232,7 +232,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         let client = BTLocalPaymentClient(authorization: tempClientToken)
         client.apiClient = mockAPIClient
         client.webAuthenticationSession = mockWebAuthenticationSession
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains(BTLocalPaymentAnalytics.paymentStarted))
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains(BTLocalPaymentAnalytics.browserPresentationSucceeded))
@@ -264,7 +264,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains(BTLocalPaymentAnalytics.browserLoginAlertCanceled))
     }
@@ -295,7 +295,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains(BTLocalPaymentAnalytics.paymentFailed))
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains(BTLocalPaymentAnalytics.browserPresentationFailed))
     }
@@ -315,7 +315,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         client.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/local-payment/success?PayerID=PCKXQCZ6J3YXU&paymentId=PAY-79C90584AX7152104LNY4OCY&token=EC-0A351828G20802249")!)
 
@@ -339,7 +339,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         client.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/local-payment/success")!)
 
@@ -360,7 +360,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { _, _ in }
+        client.start(localPaymentRequest) { _, _ in }
 
         client.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/local-payment/success")!)
 
@@ -380,7 +380,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
         
-        client.startPaymentFlow(localPaymentRequest) { _, error in
+        client.start(localPaymentRequest) { _, error in
             guard let error = error as NSError? else { return }
             XCTAssertEqual(error.domain, BTLocalPaymentError.errorDomain)
             XCTAssertEqual(error.code, BTLocalPaymentError.canceled("flow-type").errorCode)
@@ -401,7 +401,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
             ]
         )
 
-        client.startPaymentFlow(localPaymentRequest) { result, error in
+        client.start(localPaymentRequest) { result, error in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
         }
@@ -429,7 +429,7 @@ class BTLocalPaymentClient_UnitTests: XCTestCase {
         )
         paymentRequest.localPaymentFlowDelegate = mockLocalPaymentRequestDelegate
 
-        client.startPaymentFlow(paymentRequest) { _, _ in }
+        client.start(paymentRequest) { _, _ in }
 
         client.handleOpen(
             URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/local-payment/success")!